Got habit history transitioning, crashing after open
