+++ app/src/main/java/cmput301f17t01/bronzify/activities/MyHistoryActivity.java
@@ -10,25 +10,17 @@ import android.support.v4.view.ViewPager;
 import android.support.v4.widget.DrawerLayout;
 import android.support.v7.app.ActionBarDrawerToggle;
 import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
 import android.support.v7.widget.Toolbar;
+import android.view.LayoutInflater;
 import android.view.Menu;
 import android.view.MenuItem;
-
-import java.util.ArrayList;
-import java.util.Calendar;
-import java.util.Date;
-import java.util.List;
+import android.view.View;
 
 import cmput301f17t01.bronzify.R;
-import cmput301f17t01.bronzify.adapters.recyclers.MyEventAdapter;
 import cmput301f17t01.bronzify.adapters.HabitHistoryViewPagerAdapter;
 import cmput301f17t01.bronzify.controllers.NavigationController;
 import cmput301f17t01.bronzify.models.AppLocale;
-import cmput301f17t01.bronzify.models.HabitEvent;
-import cmput301f17t01.bronzify.models.HabitType;
-import cmput301f17t01.bronzify.models.User;
 
 // import android.support.design.widget.TabItem;
 
@@ -44,7 +36,6 @@ public class MyHistoryActivity extends AppCompatActivity implements NavigationVi
     private TabLayout tabLayout;
     private DrawerLayout drawer;
     private String[] pageTitle = {"Feed", "Map"};
-    private List<HabitEvent> events = new ArrayList<HabitEvent>();
 
 
     /**
@@ -114,30 +105,14 @@ public class MyHistoryActivity extends AppCompatActivity implements NavigationVi
              *
              * @param tab
              */
+
             @Override
             public void onTabReselected(TabLayout.Tab tab) {
 
             }
         });
 
-        LinearLayoutManager linearLayoutManager = new LinearLayoutManager(this);
-        /*if (linearLayoutManager.equals(null)) {
-            Log.d("MyHistoryActivity","llm is null after init");
-        }*/
-        recyclerView = findViewById(R.id.myHistoryRecycler);
-        linearLayoutManager.setOrientation(LinearLayoutManager.VERTICAL);
-      /*  if (linearLayoutManager.equals(null)) {
-            Log.d("MyHistoryActivity","llm is null after setOrientation");
-        }*/
-        recyclerView.setLayoutManager(linearLayoutManager);
-    }
 
-    @Override
-    protected void onResume() {
-        super.onResume();
-        fillEventList();
-        MyEventAdapter myEventAdapter = new MyEventAdapter(this, events);
-        recyclerView.setAdapter(myEventAdapter);
     }
 
     /**
@@ -201,35 +176,9 @@ public class MyHistoryActivity extends AppCompatActivity implements NavigationVi
         drawer.closeDrawer(GravityCompat.START);
         return true;
     }
-    private void fillEventList(){
-        User user = AppLocale.getInstance().getUser();
-        ArrayList<HabitType> habitTypes = user.getHabitTypes();
-        events.clear();
-        for(HabitType type: habitTypes){
-            ArrayList<HabitEvent> habitEvents = type.getHabitEvents();
-            for(HabitEvent event: habitEvents){
-                Date eventDate = getZeroTimeDate(event.getGoalDate());
-                Date currentDate = getZeroTimeDate(new Date());
-                int dateDiff = eventDate.compareTo(currentDate);
-                if(dateDiff <= 0){
-                    events.add(event);
-                }
-            }
-        }
+
+    public View getTabView(int position){
+        View tab = LayoutInflater.from(MyHistoryActivity.this).inflate(R.layout.habit_history_tab_feed,null);
+        return tab;
     }
-
-    // Set time to 00:00:00
-    public static Date getZeroTimeDate(Date date) {
-        Date res = date;
-        Calendar calendar = Calendar.getInstance();
-
-        calendar.setTime(date);
-        calendar.set(Calendar.HOUR_OF_DAY, 0);
-        calendar.set(Calendar.MINUTE, 0);
-        calendar.set(Calendar.SECOND, 0);
-        calendar.set(Calendar.MILLISECOND, 0);
-
-        return calendar.getTime();
-    }
-
 }
