+++ app/src/main/res/layout/content_my_history.xml
@@ -15,5 +15,13 @@
         android:layout_height="0px"
         android:layout_weight="2" />
 
+    <android.support.design.widget.TabLayout
+        android:id="@+id/tab_layout"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:layout_marginTop="?attr/actionBarSize"
+        android:layout_below="@id/toolbar"
+        android:background="@color/Amber"
+        android:theme="@style/ThemeOverlay.AppCompat.Dark.ActionBar" />
 
 </LinearLayout>
