+++ app/src/main/res/layout/event_row.xml
@@ -3,6 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     tools:ignore="ContentDescription"
     android:orientation="vertical"
+    android:id="@+id/event_row"
     android:layout_width="wrap_content"
     android:layout_height="wrap_content"
     android:background="#EEEEEE">
